explore: refactor progress classes
